refactor(checkOrder): checkOrder 에 사용되는 도메인 로직을 훅으로 분리
